+++ mods/create-new-age/changelog.md
@@ -1,8 +1,11 @@
 ------------------------------------------------------
-Version 1.1.8
+Version 1.2.0
 ------------------------------------------------------
 #### Additions
+- Added radiation effect & Geiger counter ticking - [!60 Thank you Auralyn](https://gitlab.com/antarcticgardens/create-new-age/-/merge_requests/60)
+- Reactors now explode upon overheating
 - Added CC:Tweaked compatibility
+- New Simplified Chinese translations from abandon0320
 
 ------------------------------------------------------
 Version 1.1.7c